VPC SDK Auto Released By shengkun.qxy,Version：1.60.15

+++ services/vpc/allocate_eip_address.go
@@ -98,7 +98,7 @@ type AllocateEipAddressResponse struct {
 	RequestId       string `json:"RequestId" xml:"RequestId"`
 	AllocationId    string `json:"AllocationId" xml:"AllocationId"`
 	EipAddress      string `json:"EipAddress" xml:"EipAddress"`
-	OrderId         int    `json:"OrderId" xml:"OrderId"`
+	OrderId         int64  `json:"OrderId" xml:"OrderId"`
 	ResourceGroupId string `json:"ResourceGroupId" xml:"ResourceGroupId"`
 }
