+++ services/vpc/create_vpn_pbr_route_entry.go
@@ -102,7 +102,7 @@ type CreateVpnPbrRouteEntryResponse struct {
 	OverlayMode   string `json:"OverlayMode" xml:"OverlayMode"`
 	Description   string `json:"Description" xml:"Description"`
 	State         string `json:"State" xml:"State"`
-	CreateTime    int    `json:"CreateTime" xml:"CreateTime"`
+	CreateTime    int64  `json:"CreateTime" xml:"CreateTime"`
 }
 
 // CreateCreateVpnPbrRouteEntryRequest creates a request to invoke CreateVpnPbrRouteEntry API
