+++ services/vpc/modify_customer_gateway_attribute.go
@@ -94,7 +94,7 @@ type ModifyCustomerGatewayAttributeResponse struct {
 	IpAddress         string `json:"IpAddress" xml:"IpAddress"`
 	Name              string `json:"Name" xml:"Name"`
 	Description       string `json:"Description" xml:"Description"`
-	CreateTime        int    `json:"CreateTime" xml:"CreateTime"`
+	CreateTime        int64  `json:"CreateTime" xml:"CreateTime"`
 }
 
 // CreateModifyCustomerGatewayAttributeRequest creates a request to invoke ModifyCustomerGatewayAttribute API
