+++ services/vpc/struct_virtual_border_router_for_physical_connection_type.go
@@ -18,7 +18,7 @@ package vpc
 // VirtualBorderRouterForPhysicalConnectionType is a nested struct in vpc response
 type VirtualBorderRouterForPhysicalConnectionType struct {
 	VbrId             string `json:"VbrId" xml:"VbrId"`
-	VbrOwnerUid       int    `json:"VbrOwnerUid" xml:"VbrOwnerUid"`
+	VbrOwnerUid       int64  `json:"VbrOwnerUid" xml:"VbrOwnerUid"`
 	CreationTime      string `json:"CreationTime" xml:"CreationTime"`
 	ActivationTime    string `json:"ActivationTime" xml:"ActivationTime"`
 	TerminationTime   string `json:"TerminationTime" xml:"TerminationTime"`
